second commit

+++ Assets/Scripts/Jogador.cs
@@ -10,7 +10,7 @@ public class Jogador : MonoBehaviour
     public float myMoney = 0.0f;
     public int health;
     private Animator anim;
-    public float tapValue;
+    public int tapValue;
     // Start is called before the first frame update
     void Start()
     {
